Make natural earth data optional in binary
