+++ cmds/dbmigrate_cmd.go
@@ -1,12 +1,9 @@
 package cmds
 
 import (
-	"archive/zip"
-	"bytes"
 	"context"
 	"errors"
 	"fmt"
-	"io/fs"
 	"os"
 	"strings"
 
@@ -16,16 +13,12 @@ import (
 	"github.com/golang-migrate/migrate/v4/source/iofs"
 	_ "github.com/golang-migrate/migrate/v4/source/iofs"
 	"github.com/rs/zerolog"
-	"github.com/twpayne/go-geom"
-	"github.com/twpayne/go-shapefile"
 
 	"github.com/interline-io/log"
-	"github.com/interline-io/transitland-lib/schema/ne"
 	postgressMigrations "github.com/interline-io/transitland-lib/schema/postgres"
 	"github.com/interline-io/transitland-lib/tlcli"
 	"github.com/interline-io/transitland-lib/tldb"
 	postgressAdapter "github.com/interline-io/transitland-lib/tldb/postgres"
-	"github.com/interline-io/transitland-lib/tt"
 
 	"github.com/spf13/pflag"
 )
@@ -37,7 +30,7 @@ type DBMigrateCommand struct {
 }
 
 func (cmd *DBMigrateCommand) HelpDesc() (string, string) {
-	return "Perform database migrations and load Natural Earth geographies", ""
+	return "Perform database migrations", ""
 }
 
 func (cmd *DBMigrateCommand) HelpArgs() string {
@@ -100,158 +93,12 @@ func (cmd *DBMigrateCommand) Run(ctx context.Context) error {
 		}
 		return err
 	case "natural-earth":
-		return cmd.neLoad(ctx)
+		return errors.New("natural-earth has moved to its own command; use 'transitland dbmigrate-natural-earth' instead")
 	case "down":
 		return errors.New("unsupported command")
 	}
 	return fmt.Errorf("unknown subcommand: %s", cmd.Subcommand)
 }
-
-func (cmd *DBMigrateCommand) neLoad(ctx context.Context) error {
-	var err error
-	err = fs.WalkDir(ne.EmbeddedNaturalEarthData, ".", func(path string, info fs.DirEntry, err error) error {
-		if err != nil {
-			return err
-		}
-		if !strings.HasSuffix(path, ".zip") {
-			return nil
-		}
-		neZipData, err := ne.EmbeddedNaturalEarthData.ReadFile(path)
-		if err != nil {
-			return err
-		}
-		switch path {
-		case "ne_10m_admin_1_states_provinces.zip":
-			return cmd.neProcess(ctx, neZipData, cmd.neLoadAdmins)
-		case "ne_10m_populated_places_simple.zip":
-			return cmd.neProcess(ctx, neZipData, cmd.neLoadPlaces)
-		}
-		return nil
-	})
-	return err
-}
-
-func (cmd *DBMigrateCommand) neProcess(ctx context.Context, neZipFile []byte, cb shpFileHandler) error {
-	ret := []neShape{}
-	zipReader, err := zip.NewReader(bytes.NewReader(neZipFile), int64(len(neZipFile)))
-	if err != nil {
-		return err
-	}
-	scanner, err := shapefile.NewScannerFromZipReader(zipReader, &shapefile.ReadShapefileOptions{
-		DBF: &shapefile.ReadDBFOptions{
-			SkipBrokenFields: true,
-		},
-	})
-	if err != nil {
-		return err
-	}
-	shapeFields := scanner.DBFFieldDescriptors()
-	for scanner.Next() {
-		shpGeom, _, shpRec := scanner.Scan()
-		if shpGeom == nil {
-			continue
-		}
-		t := neShape{}
-		t.attrs = map[string]string{}
-		for i, fieldDesc := range shapeFields {
-			fieldName := fieldDesc.Name
-			fieldValue := shpRec[i]
-			t.attrs[fieldName] = fmt.Sprintf("%v", fieldValue)
-		}
-		t.geometry = shpGeom.Geom
-		ret = append(ret, t)
-	}
-	return cb(ctx, cmd.Adapter, ret)
-}
-
-type shpFileHandler = func(context.Context, tldb.Adapter, []neShape) error
-
-type neShape struct {
-	geometry geom.T
-	attrs    map[string]string
-}
-
-//////////////////
-
-type neAdmin struct {
-	Name     tt.String   `db:"name"`
-	Admin    tt.String   `db:"admin"`
-	IsoName  tt.String   `db:"iso_3166_2"`
-	IsoA2    tt.String   `db:"iso_a2"`
-	Geometry tt.Geometry `db:"geometry"`
-}
-
-func (ent *neAdmin) TableName() string {
-	return "ne_10m_admin_1_states_provinces"
-}
-
-func (cmd *DBMigrateCommand) neLoadAdmins(ctx context.Context, atx tldb.Adapter, shapes []neShape) error {
-	var ents []any
-	for _, nes := range shapes {
-		ent := neAdmin{}
-		ent.Geometry = tt.NewGeometry(nes.geometry)
-		for k, val := range nes.attrs {
-			switch k {
-			case "admin":
-				ent.Admin.Set(val)
-			case "name":
-				ent.Name.Set(val)
-			case "iso_3166_2":
-				ent.IsoName.Set(val)
-			case "iso_a2":
-				ent.IsoA2.Set(val)
-			}
-		}
-		ents = append(ents, &ent)
-	}
-	log.Info().Msgf("Inserting %d admin boundaries", len(ents))
-	if _, err := cmd.Adapter.MultiInsert(ctx, ents); err != nil {
-		return err
-	}
-	return nil
-}
-
-//////////////////
-
-type nePlace struct {
-	Name     tt.String   `db:"name"`
-	Adm0Name tt.String   `db:"adm0name"`
-	Adm1Name tt.String   `db:"adm1name"`
-	IsoA2    tt.String   `db:"iso_a2"`
-	Geometry tt.Geometry `db:"geometry"`
-}
-
-func (ent *nePlace) TableName() string {
-	return "ne_10m_populated_places"
-}
-
-func (cmd *DBMigrateCommand) neLoadPlaces(ctx context.Context, atx tldb.Adapter, shapes []neShape) error {
-	var ents []any
-	for _, nes := range shapes {
-		ent := nePlace{}
-		ent.Geometry = tt.NewGeometry(nes.geometry)
-		for k, val := range nes.attrs {
-			switch k {
-			case "name":
-				ent.Name.Set(val)
-			case "adm0name":
-				ent.Adm0Name.Set(val)
-			case "adm1name":
-				ent.Adm1Name.Set(val)
-			case "iso_a2":
-				ent.IsoA2.Set(val)
-			}
-		}
-		ents = append(ents, &ent)
-	}
-	log.Info().Msgf("Inserting %d populated places", len(ents))
-	if _, err := cmd.Adapter.MultiInsert(ctx, ents); err != nil {
-		return err
-	}
-	return nil
-}
-
-/////////
 
 type migrationLogger struct {
 	log zerolog.Logger
